fix(ocr): drop two redundant usize casts so clippy -D warnings passes
Pre-existing, and not something I would normally touch. But CI gates
rust-checks behind detect-changes, so `cargo clippy -- -D warnings` only
runs when Rust files change -- and it has not run for a while. This
branch changes Rust, so the job fires and these two turn into errors.

My change is what makes the check run, so clearing it is mine to do.

CI clippy exit 0. cargo test --lib exit 0.

+++ src-tauri/src/ocr.rs
@@ -212,7 +212,7 @@ mod imp {
             unsafe { CGContext::begin_page(Some(&ctx), &raw const page_rect) };
             CGContext::draw_pdf_page(Some(&ctx), Some(&src_page));
 
-            if let Some(ocr) = pages.iter().find(|p| p.index == index as usize) {
+            if let Some(ocr) = pages.iter().find(|p| p.index == index) {
                 draw_invisible_text(&ctx, ocr);
             }
             CGContext::end_page(Some(&ctx));
@@ -310,7 +310,7 @@ mod imp {
         let document = unsafe { PDFDocument::initWithURL(PDFDocument::alloc(), &url) }
             .ok_or_else(|| "This PDF could not be opened.".to_string())?;
 
-        let page_count = unsafe { document.pageCount() } as usize;
+        let page_count = unsafe { document.pageCount() };
         if page_count == 0 {
             return Err("This PDF has no pages.".to_string());
         }
